requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.23
csc-pysparse==1.1.1.4
csc-utils==0.6.7
decorator==4.0.9
Divisi2==2.2.5
Django==1.9.5
django-recommends==0.3.1
django-registration-redux==1.4
django-rest-swagger==0.3.5
djangorestframework==3.3.3
kombu==3.0.35
networkx==1.11
numpy==1.11.0
python-dateutil==2.5.2
python-recsys==0.2
pytz==2016.3
PyYAML==3.11
scikits.learn==0.8.1
scipy==0.17.0
six==1.10.0
